rename workflows module
