+++ src/friendly_computing_machine/cli/workflow_cli.py
@@ -20,7 +20,7 @@ from friendly_computing_machine.cli.context.temporal import (
     setup_temporal,
 )
 from friendly_computing_machine.db.db import should_run_migration
-from friendly_computing_machine.workflows.worker import run_worker
+from friendly_computing_machine.temporal.worker import run_worker
 
 logger = logging.getLogger(__name__)
