+++ src/friendly_computing_machine/temporal/sample.py
