+++ src/friendly_computing_machine/temporal/util.py
